# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Datastory API specification
 * OpenAPI spec version: 0.2.0
 */
export interface HealthStatus {
  status: string;
}

export interface LoginBody {
  email: string;
  password: string;
}

export type AuthUserRole = (typeof AuthUserRole)[keyof typeof AuthUserRole];

export const AuthUserRole = {
  admin: "admin",
  brand_admin: "brand_admin",
  viewer: "viewer",
} as const;

export interface AuthUser {
  id: string;
  email: string;
  firstname: string;
  lastname: string;
  role: AuthUserRole;
  companyId?: string | null;
  franchiseId?: string | null;
}

export interface Company {
  id: string;
  name: string;
  slug: string;
  domain: string;
  hasFranchise: boolean;
  logoUrl?: string | null;
  fontRegularFilename?: string | null;
  fontBoldFilename?: string | null;
  createdAt: string;
}

export interface CreateCompanyBody {
  name: string;
  slug: string;
  domain: string;
  hasFranchise?: boolean;
  logoUrl?: string | null;
}

export interface UpdateCompanyBody {
  name?: string;
  slug?: string;
  domain?: string;
  hasFranchise?: boolean;
  logoUrl?: string | null;
}

export interface Franchise {
  id: string;
  name: string;
  companyId: string;
  code: string;
  address: string;
  zipcode: string;
  createdAt: string;
}

export interface CreateFranchiseBody {
  name: string;
  companyId: string;
  code?: string;
  address?: string;
  zipcode?: string;
}

export interface UpdateFranchiseBody {
  name?: string;
  code?: string;
  address?: string;
  zipcode?: string;
}

export interface Dashboard {
  id: string;
  name: string;
  companyId: string;
  lookerUrl: string;
  active: boolean;
  createdAt: string;
}

export interface CreateDashboardBody {
  name: string;
  companyId: string;
  lookerUrl: string;
  active?: boolean;
}

export interface UpdateDashboardBody {
  name?: string;
  lookerUrl?: string;
  active?: boolean;
}

export interface SetDashboardUsersBody {
  userIds: string[];
}

export interface SetDashboardFranchisesBody {
  franchiseIds: string[];
}

export type UserRole = (typeof UserRole)[keyof typeof UserRole];

export const UserRole = {
  admin: "admin",
  brand_admin: "brand_admin",
  viewer: "viewer",
} as const;

export interface User {
  id: string;
  email: string;
  firstname: string;
  lastname: string;
  role: UserRole;
  companyId?: string | null;
  franchiseId?: string | null;
  isActive: boolean;
  createdAt: string;
}

export type CreateUserBodyRole =
  (typeof CreateUserBodyRole)[keyof typeof CreateUserBodyRole];

export const CreateUserBodyRole = {
  admin: "admin",
  brand_admin: "brand_admin",
  viewer: "viewer",
} as const;

export interface CreateUserBody {
  email: string;
  firstname: string;
  lastname: string;
  role: CreateUserBodyRole;
  password?: string;
  companyId?: string | null;
  franchiseId?: string | null;
}

export type UpdateUserBodyRole =
  (typeof UpdateUserBodyRole)[keyof typeof UpdateUserBodyRole];

export const UpdateUserBodyRole = {
  admin: "admin",
  brand_admin: "brand_admin",
  viewer: "viewer",
} as const;

export interface UpdateUserBody {
  email?: string;
  firstname?: string;
  lastname?: string;
  role?: UpdateUserBodyRole;
  password?: string;
  companyId?: string | null;
  franchiseId?: string | null;
  isActive?: boolean;
}

export type ListFranchisesParams = {
  companyId?: string;
};

export type ListUsersParams = {
  companyId?: string;
};