# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.2.0
  description: Datastory API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login with email and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginBody"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthUser"
        "401":
          description: Invalid credentials
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current authenticated user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthUser"
        "401":
          description: Not authenticated
...
  /companies:
    get:
      operationId: listCompanies
      tags: [companies]
      summary: List all companies
      responses:
        "200":
          description: List of companies
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Company"
    post:
      operationId: createCompany
      tags: [companies]
      summary: Create a new company
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCompanyBody"
      responses:
        "201":
          description: Created company
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Company"
  /companies/{id}:
    get:
      operationId: getCompany
      tags: [companies]
      summary: Get a company by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Company
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Company"
        "404":
          description: Not found
    put:
      operationId: updateCompany
      tags: [companies]
      summary: Update a company
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateCompanyBody"
      responses:
        "200":
          description: Updated company
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Company"
    delete:
      operationId: deleteCompany
      tags: [companies]
      summary: Delete a company
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted
...
  /dashboards:
    get:
      operationId: listDashboards
      tags: [dashboards]
      summary: List dashboards accessible to current user
      responses:
        "200":
          description: List of dashboards
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Dashboard"
    post:
      operationId: createDashboard
      tags: [dashboards]
      summary: Create a dashboard
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateDashboardBody"
      responses:
        "201":
          description: Created dashboard
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Dashboard"
  /dashboards/{id}:
    get:
      operationId: getDashboard
      tags: [dashboards]
      summary: Get a dashboard by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Dashboard
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Dashboard"
        "403":
          description: Access denied
        "404":
          description: Not found
    put:
      operationId: updateDashboard
      tags: [dashboards]
      summary: Update a dashboard
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateDashboardBody"
      responses:
        "200":
          description: Updated dashboard
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Dashboard"
    delete:
      operationId: deleteDashboard
      tags: [dashboards]
      summary: Delete a dashboard
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted
  /dashboards/{id}/users:
    get:
      operationId: getDashboardUsers
      tags: [dashboards]
      summary: Get user IDs assigned to a dashboard
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: List of user IDs
          content:
            application/json:
              schema:
                type: array
                items:
                  type: string
    put:
      operationId: setDashboardUsers
      tags: [dashboards]
      summary: Set users for a dashboard
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SetDashboardUsersBody"
      responses:
        "200":
          description: Updated
  /dashboards/{id}/franchises:
    get:
      operationId: getDashboardFranchises
      tags: [dashboards]
      summary: Get franchise IDs assigned to a dashboard
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: List of franchise IDs
          content:
            application/json:
              schema:
                type: array
                items:
                  type: string
    put:
      operationId: setDashboardFranchises
      tags: [dashboards]
      summary: Set franchises for a dashboard
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SetDashboardFranchisesBody"
      responses:
        "200":
          description: Updated
...
components:
  schemas:
...
    LoginBody:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
        password:
          type: string
    AuthUser:
      type: object
      required: [id, email, firstname, lastname, role]
      properties:
        id:
          type: string
        email:
          type: string
        firstname:
          type: string
        lastname:
          type: string
        role:
          type: string
          enum: [admin, brand_admin, viewer]
        companyId:
          type: string
          nullable: true
        franchiseId:
          type: string
          nullable: true
    Company:
      type: object
      required: [id, name, slug, domain, hasFranchise, createdAt]
      properties:
        id:
          type: string
        name:
          type: string
        slug:
          type: string
        domain:
          type: string
        hasFranchise:
          type: boolean
        logoUrl:
          type: string
          nullable: true
        fontRegularFilename:
          type: string
          nullable: true
        fontBoldFilename:
          type: string
          nullable: true
        createdAt:
          type: string
    CreateCompanyBody:
      type: object
      required: [name, slug, domain]
      properties:
        name:
          type: string
        slug:
          type: string
        domain:
          type: string
        hasFranchise:
          type: boolean
        logoUrl:
          type: string
          nullable: true
    UpdateCompanyBody:
      type: object
      properties:
        name:
          type: string
        slug:
          type: string
        domain:
          type: string
        hasFranchise:
          type: boolean
        logoUrl:
          type: string
          nullable: true
...
    Dashboard:
      type: object
      required: [id, name, companyId, lookerUrl, active, createdAt]
      properties:
        id:
          type: string
        name:
          type: string
        companyId:
          type: string
        lookerUrl:
          type: string
        active:
          type: boolean
        createdAt:
          type: string
    CreateDashboardBody:
      type: object
      required: [name, companyId, lookerUrl]
      properties:
        name:
          type: string
        companyId:
          type: string
        lookerUrl:
          type: string
        active:
          type: boolean
    UpdateDashboardBody:
      type: object
      properties:
        name:
          type: string
        lookerUrl:
          type: string
        active:
          type: boolean
    SetDashboardUsersBody:
      type: object
      required: [userIds]
      properties:
        userIds:
          type: array
          items:
            type: string
    SetDashboardFranchisesBody:
      type: object
      required: [franchiseIds]
      properties:
        franchiseIds:
          type: array
          items:
            type: string
...