# lib/api-zod/src/generated/types/authUser.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Datastory API specification
 * OpenAPI spec version: 0.2.0
 */
import type { AuthUserRole } from "./authUserRole";

export interface AuthUser {
  id: string;
  email: string;
  firstname: string;
  lastname: string;
  role: AuthUserRole;
  companyId?: string | null;
  franchiseId?: string | null;
}