# lib/api-zod/src/generated/types/company.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Datastory API specification
 * OpenAPI spec version: 0.2.0
 */

export interface Company {
  id: string;
  name: string;
  slug: string;
  domain: string;
  hasFranchise: boolean;
  logoUrl?: string | null;
  fontRegularFilename?: string | null;
  fontBoldFilename?: string | null;
  createdAt: string;
}