# lib/api-zod/src/generated/types/createCompanyBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Datastory API specification
 * OpenAPI spec version: 0.2.0
 */

export interface CreateCompanyBody {
  name: string;
  slug: string;
  domain: string;
  hasFranchise?: boolean;
  logoUrl?: string | null;
}