# lib/api-zod/src/generated/types/createDashboardBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Datastory API specification
 * OpenAPI spec version: 0.2.0
 */

export interface CreateDashboardBody {
  name: string;
  companyId: string;
  lookerUrl: string;
  active?: boolean;
}