# lib/api-zod/src/generated/types/updateDashboardBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Datastory API specification
 * OpenAPI spec version: 0.2.0
 */

export interface UpdateDashboardBody {
  name?: string;
  lookerUrl?: string;
  active?: boolean;
}